graphql: foist a 'Transport' interface into play.
This introduces a little more room for flexibility than mere
http.Client replacement.  It should be useful for testing.
(Earlier in development, I had some additional Transport impls
readymade for testing; those are not included in this commit, because
they came out somewhat more opinionated than seems correct to ship
in a library like this one.  But there's some good stuff that can
be done with this interface for sure!)

Request and Response types that were previously used in the unexported
methods doing the HTTP calls are now exported so Transport can refer
to them.  This means the Transport interface's responsibilities include
both marshalling (presumably to JSON) as well as actual binary
transport.  This is a design choice I wavered on, and actually tried
several ways (earlier drafts had a Marshaller component passed in --
added complexity, didn't feel significantly better; another had
transports as plain binary pipes; this worked, but on further reading
the graphQL spec actually says JSON *isn't* strictly mandated as the
format, so binary pipes that just happened to presume it seemed
like a poor design...); long story short, this seems least wrong.

It's worth mentioning at least briefly that there's an argument for
doing transport pluggability with plain http.Client/http.RoundTripper.
This doesn't hold up well, IMO.  The semantics of HTTP are not supposed
to be relevant to graphQL, so it's immediately ideologically tainted.
The number of nested interface declarations required to implement one
of these is kind of high.  And if you ever want to make a delegate
transport which wants to examine the request produced by another?
Troubled waters; the http interfaces were designed around readers
and writers because they're supposed to be streaming, and that does not
jive smoothly with the idea of examination without mutation.  (This
is something that came up almost immediately when making tester
transport mocks, and so seems relevant.)

There's still a `json.RawMessage` field in the Response type.
This does also still restrict us to JSON serial formats in practice.
Extracting it is frankly too hard; but if anyone wants to try, PR's
welcome.  (It would get deep fast, though.  The reflective mappers
are also all tied up with json implementation details.)