transport: move check for default http client closer to site.
This makes it less surprising to use the NewPluggableClient with
a TransportHTTP as compared to doing the same thing with NewClient.

(This comes up for example when you need to configure a TransportHTTP
to use as a *delegate* even when specifying another Transport to
NewPluggableClient for e.g. testing interception purposes.)